refactor: move pickle_dump and pickle_load to utils/file;
